+++ src/apis/tokenAPI.ts
@@ -1,4 +1,4 @@
-import { authApiPost } from "./apiUtils";
+import { authApiPost } from "./axios/apiUtils";
 import { postReissueResponse } from "./types/token";
 
 export const postReissue = async (refreshToken: string) => {
